backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-bcrypt==1.0.1
python-dotenv==1.1.0
requests==2.32.5
